requirements.txt
langchain
langchain-openai
langgraph
openai
Pillow
python-dotenv
fastapi
uvicorn
sqlalchemy
pymysql
bcrypt
pyjwt
python-multipart
pydantic-settings
